+++ integration-tests/emulated/tests/assets/asset-hub-kusama/src/tests/teleport.rs
@@ -32,15 +32,15 @@ fn penpal_to_ah_foreign_assets_sender_assertions(t: ParaToSystemParaTest) {
 		PenpalA,
 		vec![
 			RuntimeEvent::ForeignAssets(
-				pallet_assets::Event::Burned { asset_id, owner, .. }
+				pallet_assets::Event::Withdrawn { asset_id, who, .. }
 			) => {
 				asset_id: *asset_id == system_para_native_asset_location,
-				owner: *owner == t.sender.account_id,
+				who: *who == t.sender.account_id,
 			},
-			RuntimeEvent::Assets(pallet_assets::Event::Burned { asset_id, owner, balance }) => {
+			RuntimeEvent::Assets(pallet_assets::Event::Withdrawn { asset_id, who, amount }) => {
 				asset_id: *asset_id == expected_asset_id,
-				owner: *owner == t.sender.account_id,
-				balance: *balance == expected_asset_amount,
+				who: *who == t.sender.account_id,
+				amount: *amount == expected_asset_amount,
 			},
 		]
 	);
@@ -54,14 +54,13 @@ fn penpal_to_ah_foreign_assets_receiver_assertions(t: ParaToSystemParaTest) {
 	assert_expected_events!(
 		AssetHubKusama,
 		vec![
-			RuntimeEvent::Balances(pallet_balances::Event::Minted { who, .. }) => {
+			RuntimeEvent::Balances(pallet_balances::Event::Deposit { who, .. }) => {
 				who: *who == t.receiver.account_id,
 			},
-			RuntimeEvent::ForeignAssets(pallet_assets::Event::Issued { owner, amount, .. }) => {
-				owner: *owner == t.receiver.account_id,
+			RuntimeEvent::ForeignAssets(pallet_assets::Event::Deposited { who, amount, .. }) => {
+				who: *who == t.receiver.account_id,
 				amount: *amount == expected_foreign_asset_amount,
 			},
-			RuntimeEvent::Balances(pallet_balances::Event::Issued { .. }) => {},
 		]
 	);
 }
@@ -75,10 +74,10 @@ fn ah_to_penpal_foreign_assets_sender_assertions(t: SystemParaToParaTest) {
 		AssetHubKusama,
 		vec![
 			// foreign asset is burned locally as part of teleportation
-			RuntimeEvent::ForeignAssets(pallet_assets::Event::Burned { asset_id, owner, balance }) => {
+			RuntimeEvent::ForeignAssets(pallet_assets::Event::Withdrawn { asset_id, who, amount }) => {
 				asset_id: *asset_id == expected_foreign_asset_id,
-				owner: *owner == t.sender.account_id,
-				balance: *balance == expected_foreign_asset_amount,
+				who: *who == t.sender.account_id,
+				amount: *amount == expected_foreign_asset_amount,
 			},
 		]
 	);
@@ -94,15 +93,15 @@ fn ah_to_penpal_foreign_assets_receiver_assertions(t: SystemParaToParaTest) {
 		PenpalA,
 		vec![
 			// local asset is teleported into account of receiver
-			RuntimeEvent::Assets(pallet_assets::Event::Issued { asset_id, owner, amount }) => {
+			RuntimeEvent::Assets(pallet_assets::Event::Deposited { asset_id, who, amount }) => {
 				asset_id: *asset_id == expected_asset_id,
-				owner: *owner == t.receiver.account_id,
+				who: *who == t.receiver.account_id,
 				amount: *amount == expected_asset_amount,
 			},
 			// native asset for fee is deposited to receiver
-			RuntimeEvent::ForeignAssets(pallet_assets::Event::Issued { asset_id, owner, .. }) => {
+			RuntimeEvent::ForeignAssets(pallet_assets::Event::Deposited { asset_id, who, .. }) => {
 				asset_id: *asset_id == Location::parent(),
-				owner: *owner == t.receiver.account_id,
+				who: *who == t.receiver.account_id,
 			},
 		]
 	);
@@ -480,7 +479,7 @@ fn relay_origin_assertions(t: RelayToSystemParaTest) {
 		Kusama,
 		vec![
 			// Amount to teleport is withdrawn from Sender
-			RuntimeEvent::Balances(pallet_balances::Event::Burned { who, amount }) => {
+			RuntimeEvent::Balances(pallet_balances::Event::Withdraw { who, amount }) => {
 				who: *who == t.sender.account_id,
 				amount: *amount == t.args.amount,
 			},
@@ -663,7 +662,7 @@ fn limited_teleport_native_assets_from_relay_to_asset_hub_checking_acc_burn_work
 					who: *who == <AssetHubKusama as AssetHubKusamaPallet>::PolkadotXcm::check_account(),
 					amount:  *amount == t.args.amount,
 				},
-				RuntimeEvent::Balances(pallet_balances::Event::Minted { who, .. }) => {
+				RuntimeEvent::Balances(pallet_balances::Event::Deposit { who, .. }) => {
 					who: *who == t.receiver.account_id,
 				},
 				RuntimeEvent::MessageQueue(
@@ -738,7 +737,7 @@ fn limited_teleport_native_assets_from_asset_hub_to_relay_checking_acc_mint_work
 			AssetHubKusama,
 			vec![
 				RuntimeEvent::Balances(
-					pallet_balances::Event::Burned { who, amount }
+					pallet_balances::Event::Withdraw { who, amount }
 				) => {
 					who: *who == t.sender.account_id,
 					amount: *amount == t.args.amount,
@@ -760,7 +759,7 @@ fn limited_teleport_native_assets_from_asset_hub_to_relay_checking_acc_mint_work
 				RuntimeEvent::MessageQueue(
 					pallet_message_queue::Event::Processed { success: true, .. }
 				) => {},
-				RuntimeEvent::Balances(pallet_balances::Event::Minted { who, .. }) => {
+				RuntimeEvent::Balances(pallet_balances::Event::Deposit { who, .. }) => {
 					who: *who == t.receiver.account_id,
 				},
 			]